cleanup
consistently use uint16_t for rpm, was using float in some places

+++ CoreProxy.h
@@ -35,7 +35,7 @@
 class CoreProxy : public Core
 {
 private:
-    float rpm;
+    uint16_t rpm;
     bool isAlarm;
     bool powerOn;
     bool isPanic;
